Fixato il problema su modifica textView del fragment da mainActivity
Ricorda: quando si committa un fragment sull’interfaccia, non verrà
immediatamente inserito! Quindi tutti i gestori di eventi del fragment
come onCreateView oppure onActivityCreated e simili non verranno
richiamati immediatamente, e di conseguenza non sará possibile trovare
i View del fragment immediatamente fino a che il fragment non è
effettivamente stato attaccato alla Activity.

+++ app/src/main/java/micc/beaconav/indoorEngine/IndoorMap.java
@@ -32,7 +32,7 @@ public class IndoorMap
         paint.setStyle(Paint.Style.FILL);
         paint.setColor(Color.RED);
 
-        Bitmap tempBmp =  Bitmap.createBitmap(building.getWidth(), building.getHeight(), Bitmap.Config.RGB_565);
+        Bitmap tempBmp =  Bitmap.createBitmap((int)building.getWidth(), (int)building.getHeight(), Bitmap.Config.RGB_565);
         Canvas tempCanvas = new Canvas(tempBmp);
         building.draw(tempCanvas);
